Design Patterns 8/src/edu/umb/cs680/hw08/Link.java
package edu.umb.cs680.hw08;

import java.time.LocalDateTime;

public class Link extends FSElement {

    private FSElement target;

    public Link(Directory parent, String name, LocalDateTime creationTime, FSElement target) {
        super(parent, name, 0, creationTime);
        this.target = target;
    }
    public FSElement getTarget() {
        return this.target;
    }
    public boolean isDirectory() {
        return false;
    }
    public boolean isFile() {
        return false;
    }

    public boolean isLink() {
        return true;
    }

}